Comment out parsing embedded Go source files for now
Till this point we were parsing all Go source files based on
src.GoSourceFiles to create AST on package meta initialization. It was
fine to use as in this project or Go module, but it's really annoying
when user would use sched or exec package in theirs project.

This idea will be revisited for sure. I still want to parse task's
Execute method source code, to detects changes in DAG implementations.
But we need to do it more flexible regarding how those libraries will be
used by users.