Pull logs related queries into a separate client - db.LogsClient
I don't really know how logs-related queries got under main db.Client.
Especially when those are two different databases with different
schemas... Fortunetly it's just a couple of methods which are used in
few places, so it was relatively easy to refactor.

So now we have two main clients in db package:
    - db.Client - for the main scheduler database
    - db.LogsClient - for task logs database